refactor: Repository 추상화 + DI 적용
